Created more modularity and CustomPlayer is now finished
Added a HUDDrawable interface defining whether the mod has custom HUD elements.
Updated ClientTickable.java to be client-side only.
Updated ClientEnvironment to actually be useful.
Created a PacketCreateClientSidePlayer packet which tells the ClientEnvironment to create a new player for every ClientTickable mod.
Added the entirety of the celtest package to this commit so others can test as well.
Updated Syncable.java to actually include functionality (used in many classes like CustomPlayerRegistry)

# Celtest package:
Created SanityCommand to test the persistence of CustomData.
Created SanityPlayer which is an all-in-one test class for CustomPlayer, which tests all modules:
Note: all modules are tested by Mana.java. The sanity part is merely left-over copy-pasted code from the HumanFlesh project.